Added PROIEL data, TAGSET.xml, refresh xquery files
Added a repository from the "Syntacticus-Treebanks" with the PROIEL data into a folder called PROIEL-DATA yesterday, and the TAGSET.xml file necessary for the LDT tags (retrieved from one of Perseids' GitHub repositories), and am saving my progress on the agldt_search.xqm and its companion (see the agldt_search comments and such for details on changes). Also saved the "codebook" here.